+++ planck-cljs/src/planck/io.cljs
@@ -598,11 +598,6 @@
                         :writer #(implements? IWriter %)))
   :ret boolean?)
 
-;; These have been moved
-(def ^:deprecated ^:no-doc read-line planck.core/read-line)
-(def ^:deprecated ^:no-doc slurp planck.core/slurp)
-(def ^:deprecated ^:no-doc spit planck.core/spit)
-
 (set! planck.core/*reader-fn* reader)
 (set! planck.core/*writer-fn* writer)
 (set! planck.core/*as-file-fn* as-file)